Clean up comments TODO: Add github ticket number to remaining TODOs
