+++ packages/plexus/src/Digraph/MeasurableNode.tsx
@@ -27,45 +27,11 @@ type TProps<T = {}> = Omit<TMeasurableNodeRenderer<T>, 'measurable' | 'measureNo
   vertex: TVertex<T>;
 };
 
-/*
-type TState = {
-  wasHidden: boolean;
-}
- */
-
 const SVG_HIDDEN_STYLE = { visibility: 'hidden' };
 
-export default class MeasurableNode<T = {}> extends React.PureComponent<TProps<T> /*, TState */> {
+export default class MeasurableNode<T = {}> extends React.PureComponent<TProps<T>> {
   htmlRef: React.RefObject<HTMLDivElement> = React.createRef();
   svgRef: React.RefObject<SVGGElement> = React.createRef();
-
-  /*
-  state: TState = {
-    wasHidden: false,
-  }
-   */
-  wasHidden = false;
-
-  /*
-  getDerivedStateFromProps(props: TProps<T>) {
-    if (this.props.hidden && !props.hidden) return { wasHidden: true }
-    return { wasHidden: false };
-  }
-   */
-
-  /*
-  shouldComponentUpdate(nextProps: TProps<T>) {
-    let rv = false;
-    Object.keys(nextProps).forEach((key) => {
-      const k = key as keyof TProps<T>;
-      if (nextProps[k] !== this.props[k]) {
-        rv = true;
-        console.log(`${k} changed in measurable node singular props`);
-      }
-    });
-    return rv;
-  }
-   */
 
   private measureHtml() {
     const { current } = this.htmlRef;
@@ -100,13 +66,11 @@ export default class MeasurableNode<T = {}> extends React.PureComponent<TProps<T
           position: 'absolute',
           transform:
             left == null || top == null ? undefined : `translate(${left.toFixed()}px,${top.toFixed()}px)`,
-          // transition: !this.wasHidden ? 'transform 2s' : undefined,
           visibility: hidden ? 'hidden' : undefined,
         },
       },
       getProps(setOnNode, vertex, renderUtils, layoutVertex)
     );
-    this.wasHidden = hidden;
     return (
       <div ref={this.htmlRef} {...props}>
         {renderNode(vertex, renderUtils, layoutVertex)}
@@ -121,9 +85,7 @@ export default class MeasurableNode<T = {}> extends React.PureComponent<TProps<T
       {
         className: getClassName('MeasurableSvgNode'),
         transform: left == null || top == null ? undefined : `translate(${left.toFixed()}, ${top.toFixed()})`,
-        style: hidden ? SVG_HIDDEN_STYLE : {
-          /* transition: 'transform 2s', */
-        },
+        style: hidden ? SVG_HIDDEN_STYLE : {},
       },
       getProps(setOnNode, vertex, renderUtils, layoutVertex)
     );
@@ -147,7 +109,6 @@ export default class MeasurableNode<T = {}> extends React.PureComponent<TProps<T
 
   render() {
     const { layerType } = this.props;
-    // console.log('measurable node single render');
     if (layerType === ELayerType.Html) {
       return this.renderHtml();
     }
